+++ server.js
@@ -21,8 +21,8 @@ const UserPrivilegesRoute = require("./routes/userPrivileges")
 const SemesterRoute = require("./routes/semester")
 const CashierRoute = require("./routes/cashier")
 const StudentLevelRoute = require("./routes/studentLevel")
-const CollegeLevelRoute = require("./routes/collegeLevel")
-const ShsLevelRoute = require("./routes/shsLevel")
+const SubjectsRoute = require("./routes/subjects")
+const StudentTypeRoute = require("./routes/studentType")
 require("./services/passport");
 const { graphqlHTTP } = require('express-graphql')
 const app = express();
@@ -70,9 +70,8 @@ app.use("/api/userPrivileges",UserPrivilegesRoute);
 app.use("/api/semester",SemesterRoute);
 app.use("/api/cashier",CashierRoute);
 app.use("/api/studentLevel",StudentLevelRoute);
-app.use("/api/collegeLevel",CollegeLevelRoute);
-app.use("/api/shsLevel",ShsLevelRoute);
-
+app.use("/api/subjects",SubjectsRoute);
+app.use("/api/studentType",StudentTypeRoute);
 
 
 // GRAPHQL